Updates for Windows for UError -> SError change. const change in CThread. const change in CString. Additions to make Visual Studio C++ compiler stop spewing irrelevant warnings.

+++ Source/Media/Media Sources/CWAVEMediaSource.cpp
@@ -94,7 +94,7 @@ OI<SError> CWAVEMediaSource::loadTracks(const CByteParceller& byteParceller)
 						audioStorageFormat =
 								OI<SAudioStorageFormat>(
 										SAudioStorageFormat(CDVIIntelIMAADPCMAudioCodec::mID, 16,
-												waveFormat.getNativeSamplesPerSec(),
+												(Float32) waveFormat.getNativeSamplesPerSec(),
 												AUDIOCHANNELMAP_FORUNKNOWN(waveFormat.getNativeChannels())));
 						break;
